split files; refactor createconfig prompting
